Put parameter checking into separate module
