+++ tests/test_registry.py
@@ -577,70 +577,6 @@ class TestDispatch(object):
         assert registry.dispatch(fake_request) is None
 
 
-class TestValidateParams(object):
-    def test_list(self):
-        def foo(a, b, c="baz"):
-            pass
-
-        with pytest.raises(InvalidParamsError):
-            Registry._validate_params_match(foo, ["foo"])
-        Registry._validate_params_match(foo, ["foo", "bar"])
-
-        with pytest.raises(InvalidParamsError):
-            Registry._validate_params_match(foo, ["foo", "bar", "bop", 42])
-
-    def test_varargs(self):
-        def foo(a, b="foo", *varargs):
-            pass
-
-        with pytest.raises(InvalidParamsError):
-            Registry._validate_params_match(foo, [])
-        Registry._validate_params_match(foo, ["foo", "bar"])
-        Registry._validate_params_match(foo, ["foo", "bar", 42])
-
-    def test_dict(self):
-        def foo(a, b, c="baz"):
-            pass
-
-        with pytest.raises(InvalidParamsError):
-            Registry._validate_params_match(foo, {"a": "foo"})
-        Registry._validate_params_match(foo, {"a": "foo", "b": "bar"})
-
-        with pytest.raises(InvalidParamsError):
-            Registry._validate_params_match(foo, {"a": "foo", "b": "bar", "c": "bop", "d": 42})
-
-        with pytest.raises(InvalidParamsError):
-            Registry._validate_params_match(foo, {"a": "foo", "c": "bar"})
-
-    def test_kwargs(self):
-        def foo(a, b, c="baz", **kwargs):
-            pass
-
-        with pytest.raises(InvalidParamsError):
-            Registry._validate_params_match(foo, {"a": "foo"})
-        Registry._validate_params_match(foo, {"a": "foo", "b": "bar"})
-
-        Registry._validate_params_match(foo, {"a": "foo", "b": "bar", "d": 42})
-
-        with pytest.raises(InvalidParamsError):
-            Registry._validate_params_match(foo, {"a": "foo", "c": "bar"})
-
-    def test_no_defaults(self):
-        def foo(a):
-            pass
-        with pytest.raises(InvalidParamsError):
-            Registry._validate_params_match(foo, [])
-        Registry._validate_params_match(foo, ["bar"])
-        with pytest.raises(InvalidParamsError):
-            Registry._validate_params_match(foo, ["bar", "baz"])
-
-        with pytest.raises(InvalidParamsError):
-            Registry._validate_params_match(foo, {})
-        Registry._validate_params_match(foo, {"a": "bar"})
-        with pytest.raises(InvalidParamsError):
-            Registry._validate_params_match(foo, {"a": "bar", "b": "baz"})
-
-
 def test_describe():
     registry = Registry()
